Rename dmv to identity-manager-sts
In order to keep naming for Infratographer projects mostly consistent,
dmv is being renamed to identity-manager-sts (for Security Token
Service, per RFC 8693). This commit does all the necessary renaming,
and also trims some incorrect documentation from the README.
