+++ README.md
@@ -1,6 +1,6 @@
-# dmv - please proceed to the next kiosk
+# identity-manager-sts - please proceed to the next kiosk
 
-dmv is an OAuth service that supports the following grant types:
+identity-manager-sts is an OAuth service that supports the following grant types:
 
 * [RFC 8693][rfc8693]: Token Exchange
 
@@ -8,7 +8,7 @@ dmv is an OAuth service that supports the following grant types:
 
 ## Usage
 
-dmv is a Go service. To build it, you can either use `make build` to build a Go binary or `make up` to both build and start the service.
+identity-manager-sts is a Go service. To build it, you can either use `make build` to build a Go binary or `make up` to both build and start the service.
 
 The `up` Makefile target will auto-generate a private key and mount it in the container for testing purposes. Note that this is not recommended for actual production use, and is merely a handy feature to allow developers to test.
 
@@ -21,7 +21,7 @@ $ read -s -p 'Enter your token: ' AUTH_TOKEN && echo
 $ curl --user my-client:foobar -XPOST -d "grant_type=urn:ietf:params:oauth:grant-type:token-exchange&subject_token=$AUTH_TOKEN&subject_token_type=urn:ietf:params:oauth:token-type:jwt" http://localhost:8000/token | jq
 ```
 
-This sends an RFC 8693 token exchange request to dmv, which then will validate the given subject token against the configured subject token issuers. If the token is valid, you will receive a response like so (access token truncated for brevity):
+This sends an RFC 8693 token exchange request to identity-manager-sts, which then will validate the given subject token against the configured subject token issuers. If the token is valid, you will receive a response like so (access token truncated for brevity):
 
 ```
 {
@@ -40,7 +40,7 @@ $ echo "$ACCESS_TOKEN" | jq '.access_token | split(".") | .[1] | @base64d | from
   "client_id": "my-client",
   "exp": 1670354213,
   "iat": 1670354113,
-  "iss": "https://dmv.infratographer.com/",
+  "iss": "https://iam.infratographer.com/",
   "jti": "e36322d3-414c-4da2-91a8-f19a6e9fb1d3",
   "scp": [],
   "sub": "my-user-id"
@@ -49,32 +49,15 @@ $ echo "$ACCESS_TOKEN" | jq '.access_token | split(".") | .[1] | @base64d | from
 
 [jq]: https://stedolan.github.io/jq/
 
-### Claim mapping
-
-DMV supports mapping of subject token claims to issued token claims using [CEL][cel]. `oauth.claimMappings` in the config file defines the mappings of issued token claims to CEL expressions. For example, the following config snippet will map the `infratographer:group` claim based on the `sub` claim:
-
-```yaml
-oauth:
-  claimMappings:
-    "infratographer:group": "claims.sub == '1234' ? 'admin' : 'user'"
-```
-
-The following variables are predefined in the CEL runtime environment:
-
-* `claims`: A dynamic map containing all claims in the subject token JWT. Accessible using `claims.*` (e.g., `claims.sub`)
-* `subSHA256`: The hex-encoded SHA256 sum of the `sub` claim
-
-[cel]: https://github.com/google/cel-go
-
 ### JWKS
 
-The [JSON Web Key Set][jwks] (JWKS) used for signing dmv JWTs is available at `/jwks.json`.
+The [JSON Web Key Set][jwks] (JWKS) used for signing identity-manager-sts JWTs is available at `/jwks.json`.
 
 [jwks]: https://www.rfc-editor.org/rfc/rfc7517.html#section-5
 
 ### Configuration
 
-dmv requires a configuration file to run. An example can be found at `dmv.example.yaml`.
+identity-manager-sts requires a configuration file to run. An example can be found at `sts.example.yaml`.
 
 Private keys must be explicitly configured with a JWT signing algorithm, such as HS256 or RS256. Symmetric keys are loaded from key files as raw bytes. All asymmetric (i.e., RSA) signing keys must be encoded using [PKCS #8][pkcs8]. To generate an RSA private key for development, the following command should get you started:
 
@@ -88,7 +71,7 @@ Update the config file and/or Docker Compose volume mounts accordingly.
 
 ## Development
 
-DMV includes a [dev container][dev-container] for facilitating service development. Using the dev container is not required, but provides a consistent environment for all contributors as well as a few perks like:
+identity-manager-sts includes a [dev container][dev-container] for facilitating service development. Using the dev container is not required, but provides a consistent environment for all contributors as well as a few perks like:
 
 * [gopls][gopls] integration out of the box
 * Host SSH auth socket mount
@@ -108,7 +91,7 @@ The provided dev container listens for SSH connections on port 2222 and bind mou
 To do so, update your `~/.ssh/config` to support agent forwarding. The following config snippet should accomplish this for you:
 
 ```
-Host dmv-devcontainer
+Host identity-manager-sts-devcontainer
   ProxyJump YOUR_HOST_HERE
   Port 2222
   User vscode
