+++ pkg/api/v1/doc.go
@@ -1,2 +1,2 @@
-// Package v1 contains top-level types and functions for DMV.
+// Package v1 contains top-level types and functions for identity-manager-sts.
 package v1
