+++ pkg/routes/doc.go
@@ -1,2 +1,2 @@
-// Package routes contains route handlers and logic for DMV.
+// Package routes contains route handlers and logic for identity-manager-sts.
 package routes
