+++ pkg/routes/routes.go
@@ -6,7 +6,7 @@ import (
 	"github.com/ory/fosite"
 	"go.uber.org/zap"
 
-	"go.infratographer.com/dmv/pkg/fositex"
+	"go.infratographer.com/identity-manager-sts/pkg/fositex"
 )
 
 // Router is the router for the application.
