+++ pkg/storage/config.go
@@ -1,6 +1,6 @@
 package storage
 
-// Config represents the storage configuration for DMV.
+// Config represents the storage configuration for identity-manager-sts.
 type Config struct {
 	Type     EngineType
 	SeedData SeedData
