+++ pkg/storage/doc.go
@@ -1,2 +1,2 @@
-// Package storage contains types and functions for managing data storage in DMV.
+// Package storage contains types and functions for managing data storage in identity-manager-sts.
 package storage
