+++ sts.example.yaml
